+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- converting function on index page
 - editing by default
 - change reference source
 - filename in editor page title
@@ -10,7 +11,6 @@
 - glang: edit uploaded files in onlyoffice editor
 - golang: delete files on the server
 - golang: show files history
-- golang: converting function on index page
 - php-laravel: new integration example
 - php-laravel: upload files to the server
 - php-laravel: create blank files and files with sample content